Step 1: decompose _command_classification.py (#4665)
Extract the GitHub mutation cardinality/route analysis from
_command_classification.py into a new sibling module
_github_mutation_analysis.py. The new module is stdlib-only and shares
the same hooks/ bootstrap.

Cross-module references (command_verb_and_args, _tokenize_command_segments_with_redirects,
_normalize_executable, _partition_output_redirects,
_extract_interpreter_segment_specs, _segment_evaluates_shell_payload,
extract_shell_command_payloads) are resolved lazily via function-local
imports to avoid the circular dependency at module load.

The facade re-exports GitHubMutationStatus, GitHubMutationKind,
GitHubMutationRecord, GitHubMutationAnalysis, and analyze_github_mutations
via the three-way discriminator pattern. The constant
_GH_READ_ONLY_SUBCOMMANDS is re-bound at module level so existing
monkeypatch.setattr sites in tests continue to work.
